feat: add esm.mjs file
